+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 INTERFACES CHANGELOG
 --------------------
 
+* [FEATURE] add AccountSwitcher with PCO-API backend
+  + update `interfaces_header` with injection of Person and Organization
+    - require current Person as `person`
+    - require current Organization as `organization`
+    - require a person's profile path as `person_profile_path`
+* [DEPENDENCY] add react-rails
+  + inject harmony config
+* [BREAKING CHANGE] apps must supply Underscore/Lodash implementation
+  + add lodash to Dummy app
+  + to update simply include Underscore/Lodash, required before interfaces
 * [BREAKING CHANGE] update hard-coded links from /styleguide/* -> /interfaces/*
 * [FEATURE] add /engine/version route for `pco interfaces` CLI
   * Interfaces now constrains the routes based on environment.